requirements.txt
streamlit
beautifulsoup4
google-generativeai
requests
python-dotenv
